+++ src/TestExplorer/Helpers/testResult.ts
@@ -2,5 +2,6 @@ export default interface TestResult {
   output?: string;
   message?: string;
   success?: boolean;
+  skipped?: boolean;
   line?: number;
 }